TST: create: Add plain prefix to OBSCURE_FILENAME
Otherwise the 'git submodule add' call will fail in the
DATALAD_TESTS_OBSCURE_PREFIX=- test case (see [0]).  Our
add_submodule() should probably abort if the path starts with "-"
because the error by 'git submodule' is very misleading, and the issue
should arguably be fixed upstream, but in either case, this doesn't
have anything to do with this series, so work around this issue in the
test.

[0]: https://travis-ci.org/datalad/datalad/builds/520413270