BF/ENH: create-sibling with --ui true for publish_on_github cast
In #3318 (a4d40a607cb54a5cd4b49459496c6c64449ec666 in particular) I changed behavior
that no post-update hook  is installed by default.  And that is indeed desired since
unless it is a publicly facing via http git repository there is no need in such a hook.
BUT ATM the hook is unconfigurably melds two effects together:

  - making this repo accessible via dummy git http backend (by running git update-server-info)
  - and generating web listing metadata

There is no configuration to enable one but not another action in the hook.

So for this cast to still work with current functionality, we would need to enable web UI
which will establish the hook (and also install web UI index.html and all the needed JS
libraries).  In the long run I guess we should disentangle two aspects somehow and
grow the hook with needed feature depending on the configuration.