TST(BK): demonstrate the regression from 0.11.4-5-g78e00dcd - remote submodule is not "initialized"
Recently 0.11.4-5-g78e00dcd removed initialization of git submodules on remote
within .git/hooks since it was indiscriminate to the cases where such
initialization was not even desired.  Initially such initialization as was introduced
because our submodule listing was using "git submodule" directly and now we do
parsing ourselves and probably because of that probably it all works even
without proper "git submodule update --init  newsubmodule" to make it properly
registered within .git/config.

This test is primarily to demonstrate the effect and seek decision on either it should
be addressed or not.  The main problematic use case I see which might mandate
action is discovery of the web ui not being populated properly, especially whenever
remote host has somewhat datalad DataLad