ENH: short -R for --recursion-limit
Because of otherwise too long of a run time I often need to use --recursion-limit
for operations such as diff while working with deep hierarchies. So I felt that it
would be nice to have a short cmdline option for it.  Another to  -R was -L (tree
uses it), but we do use -L already in elderly  datalad ls   so decided to go for -R