ENH(TST): test_integration1 to do regression testing of github interactions using VCR
This is more of a prototype, which somewhat echoes the setup (I think) used by
PyGithub itself.  They pickle interactions for their tests.  I guess whenever Github
API or pygithub behavior changes, this VCR tape would no longer be valid, but that is
yet to be seen.

To produce this tape I generated new token on github, used it within my local account
configuration, then sanitized produced tape a bit to remove it

  sed -E "s,74463[^ ]{35},TOKEN,g" -i datalad/distribution/tests/vcr_cassettes/github_yarikoptic.yaml

and then removed it from github anyways, so there should be no authorization
detail leak in the VCR tape.  In the long run, if we decided to follow this setup
it all could be streamlined more