ENH: --cfg option for rev-create to specify cfg_ procedure(s) to run
ATM there is no easy way to specify procedures to run on a freshly
created dataset. But in my case it is a common use case to quickly
create datasets where I know what type they should be - should not
place text files to git, or might be BIDS datasets, etc.  So it would
be nice to be able to specify succinctly procedure(s) to be ran on
the new dataset.

Alternative implementation could adding  --proc-post to the rev-create
command itself.

Pros would be:

  - consistent with --proc-post for datalad command itself
  - could then run any procedure, not only cfg_ one

Cons:

  - would be notably more long/tedius to type - i.e.

    rev-create --proc-post cfg_text2git

    vs

    rev-create -c text2git