ENH: AnnexRepo.get_metadata(batch=True) (fixes gh-2165)
This is a fragile enterprise due to the limitations of
`BatchedCommand()`. When queried for a non-annexed file, git annex
will error not via JSON response, but through stderr

```
{"file": "."}
git-annex: not an annexed file: .
```

which ATM is not monitored at all. I assume due to the general
sadness of subprocess under PY2.

Docstrings warn about this, there is no `batch=None` automagic because
of this.

Other than that, is works.