RF: annexrepo: Tighten decision to commit with temporary index
We decide to commit with a temporary index if there are any files with
staged changes that should _not_ be committed or if any of the
specified paths have unstaged changes.  The latter actually doesn't
require a temporary index; we just need to update the paths in the
current index, as we already do.

As a side note, by adding the files, gh-3009 introduced a subtle
change in behavior.  Before the specified files would be added to git
by 'git commit'.  After gh-3009, they are added with 'git annex add'.
This is unlikely to matter because the behavior was undefined and is
only visible when calling AnnexRepo.commit() directly as save adds the
paths before calling .commit().