RF: Aborb rev-diff
Replace core's diff with -revolution's.  Update the few diff callers
to use the new variant, with the exception of annotate_paths.  Keep
the older diff around but hidden (not listed in an interface group and
named with an underscore) for annotate_paths.  We could update it to
use the new diff, but it's not worth the effort because the plan is to
retire annotate_paths.

This is copied from revolution's 38df2ddd2858fa8786b6fcca6c472ef9be8a53e8.
You can verify that there are no substantial changes by checking
datalad-revolution out at that commit, pointing $DL_REV_DIR to the
repository, and then running

  % git diff --no-index -- \
    $DL_REV_DIR/datalad_revolution/revdiff.py \
    datalad/core/local/diff.py

  % git diff --no-index -- \
    $DL_REV_DIR/datalad_revolution/tests/test_diff.py \
    datalad/core/local/tests/test_diff.py

Closes #3190.