TST: add: Avoid create(..., save=False)
These tests rely on create's save argument, which rev_create dropped.
In test_add_dirty_tree, the save=False is unnecessary, so use a plain
rev_create call.  In the other spots, use an unbound rev_create call
to avoid saving the subdataset in the superdataset.  The results
aren't identical, but the differences don't matter in the context of
these tests.