MNT: Drop rev_run interface
eba3ccfd6 (BF: run: Save results with rev-save, 2019-03-18) added this
for compatibility, but it would have made more sense to add it to the
revolution extension, as is done for all the other rev- commands.  Do
that now.

This also frees up this module for the main Run interface class, which
should eventually live in core/local/ instead of interface/.