MNT: Avoid invalid escape sequences in strings
For a good number of regexps and a few other strings, we don't use an
r-prefix when the string contains non-escape-sequence backslashes.
This doesn't make a functional difference at the moment because, when
Python encounters a backslash that produces an invalid escape
sequence, it automatically escapes the backslash.  But, with the goal
of eventually making invalid escape sequences a syntax error, Python
v3.6 produces a DeprecationWarning, and Python 3.8 produces a
SyntaxWarning [0], so we should adjust these cases to use an r-prefix.

The instances fixed in this commit were found by the py3.8 Travis run
from gh-3406 [1].

[0]: https://bugs.python.org/issue32912
[1]: https://travis-ci.org/datalad/datalad/jobs/532901906#L1891

Closes #3418.