BF+RF: unlock: Rewrite using status()
If a file without content is passed to unlock(), it yields a
result (with a status of "impossible") saying that the file doesn't
have content.  But if the files without content weren't explicitly
provided as paths (e.g., no path was given or a directory containing
the files was given as the path), then files without content lead to a
CommandError from the underlying 'git annex unlock' call.  (Or in v6+,
they lead to the pointer being unlocked which is inconsistent with the
error given when a file is explicitly specified in v6+ and also
potentially confusing to callers.)

To fix the "no paths specified, missing content" behavior, rely on the
information provided by status() to restrict the unlock calls to files
with content.  This fix raises the question of how to report files
without content when the path isn't explicitly given.  Because
'datalad unlock' at the dataset or a directory level could potentially
lead to a lot of results, avoid reporting "can't unlock" results when
explicit paths aren't given.  Doing this at the __call__ level rather
than as a result filter means we can do the faster status(...,
untracked=no") query unless explicit, non-directory paths are given.

Aside from the fix above, this tries to match the behavior of the
previous unlock().  The main behavioral change, shown by the test
updates, is that unlock() no longer infers the dataset from a path
when the current working directory is outside a dataset and no dataset
is explicitly given.  This behavioral change is consistent with, for
example, distribution.add vs core.local.save.

This nudges annotate_paths() closer to retirement (gh-3368).