BF: interface.utils: Make default renderer handle pathlib objects
If the default renderer _were_ to run with the status results, it
would fail with

  relpath() argument must be str or bytes, not 'PosixPath'
  [genericpath.py:_check_arg_types:143] (TypeError)

This codepath isn't accessible when calling the status() command
because status() uses a tailored renderer as its custom renderer (see
gh-2481).  However, if another command that uses the default renderer
runs status() with on_failure="ignore" and then yields the result, as
unlock() will do in the next commit, then we will hit the above
failure.