OPT: gitrepo.dirty: Use 'git status --porcelain'
While in general it makes sense to use status() as the uniform
interface for querying the repository state, .dirty doesn't benefit
from it.  It takes no parameters and returns a boolean, so using git
status to answer the question is simple and less expensive.

Re: https://github.com/datalad/datalad/issues/3342#issuecomment-495781103