DOC: run/rerun: Tell caller how to inspect dirty repo
Some callers may not know how to find the unsaved modifications, so
direct them to status.  Of course they also might not know how to deal
with those modifications, but, in the interest of keeping this message
to a somewhat reasonable length, let's not take responsibility for
directing the caller to `datalad save` as one of the many ways to
resolve a dirty state.  We could also mention --explicit, but that
would make the message longer, add one more thing for a caller to be
confused about, and mention an approach that at least I don't want to
encourage.

Re: https://github.com/datalad/datalad/issues/3342#issuecomment-495974665