BF: create: Restore saving behavior for --dataset
As of fc25ac55d (BF: Do not duplicate/auto-convert dataset args to
instances, 2019-05-20), EnsureDataset returns a string rather than a
dataset when given a string.  This introduced a regression in creating
and then saving a subdataset to the parent dataset from the command
line (e.g., `datalad create -d. subds`) because downstream code
expects the dataset to be a dataset instance.  Update that code to use
refds, which is the Dataset instance for the dataset argument.

Fixes #3463.