BF: save: Honor on_failure
save() calls status() with the default on_failure value of "continue",
which means that---regardless of the on_failure value specified for
save()---status() always raises an IncompleteResultsError exception
when a failure occurs and then save() aborts.  Make save() call
status() with on_failure="ignore" so that the *save* caller gets to
decide what to do in the case of a failure.

Note that this results in some double reporting:

    [ERROR  ] path not underneath this dataset [status(...)]
    status(error): ../a [path not underneath this dataset]

But if save() were to just swallow the result so that only the log
message is shown, an on_failure value of "continue" or "stop" would
not raise an IncompleteResultsError exception or exit on the command
line with a non-zero status.