BF: download_url: Update for new path resolution logic
download_url() needs to be updated in two ways for the new path
resolution logic from gh-3435: (1) the dataset argument is assumed to
be a dataset instance if it is specified but as of fc25ac55d
EnsureDataset no longer converts a string dataset to an instance
and (2) paths are to be taken as relative to the current working
directory unless a dataset _instance_ is given.

We could update get_dataset_pwds(), but let's instead stop using it in
download_url() because it was used to avoid duplicating some logic
and, with the new path resolution, most of that logic has fallen away.
Also, the fate of get_dataset_pwds() isn't clear because its other
caller, run(), also needs to be updated for gh-3435 but that is likely
to be more involved.

Fixes #3468.