DOC+TST: save: Update for "no commits" sub-repo fix in Git
Saving an untracked sub-repository without a commit checked out [0]
results in an ugly state: the sub-repository is _not_ added as a
submodule but its untracked files are added as blobs in the
_superdataset_.  This is not usually an issue for datasets because
these get initial commits on creation, so we documented this as a
known issue rather than introducing additional checks that would come
with a performance penalty.

The core problem is a combination of 'git submodule add' accepting a
sub-repository on an unborn branch and 'git ls-files' considering such
a sub-repository a directory rather than a repository.  Both issues
are fixed in Git 2.22.0 [1].  Update the documentation note and tests
accordingly.

[0] Usually this is just a repository without any commits, but it
    could be repository that has commits but is currently on an unborn
    branch.
[1] https://public-inbox.org/git/20190409230737.26809-1-kyle@kyleam.com/
    Commits e13811189b and b22827045e.

Re: #3139