TST: utils: Add known_failure_appveyor decorator
We already have a few places where we skip a test on AppVeyor due to a
known failure.  Add a dedicated known failure decorator for these.

In many cases, this should be preferable to known_failure_windows(),
as DataLad developers primarily use GNU/Linux and we know only that a
failure happens on AppVeyor, not more broadly on Windows.