BF: create: Explicitly convert the dataset argument to instance
The path handling in create() relies on the dataset argument coming in
as a dataset instance or None.  In particular, path is expected to be
an absolute path, but this is not the case if dataset is given as a
string.  This problem is more likely to be encountered since
fc25ac55d (BF: Do not duplicate/auto-convert dataset args to
instances, 2019-05-20) because the dataset argument is no longer
automatically converted to a dataset instance.  Before this, the issue
could only be trigger by calling the unbound dataset method with a
string as the dataset.

Fixes #3484.