BF: gitrepo: Don't pass added submodules to git{,-annex} add
If we see an untracked directory in the get_content_info(...,
untracked="all") output, we know it's a sub-repository because
otherwise git would show the untracked files individually (or nothing
at all in the case of an empty directory).

We call add_submodule on these paths, which takes care of adding the
paths to the index, there's no reason to later pass these paths to
'git add' or 'git annex add'.

Most of the time that is just a noop, but it is problematic in one
situation, stemming from a change in how Git treats sub-repositories
without any commits checked out.  Starting with v2.22.0, these are
treated as any other repository, so the call to add_submodule will
appropriately fail.  Before v2.22.0, untracked files in the
sub-repository would confusingly be added to the _parent_ repo.  The
problems comes when the all of the following conditions are met:
DATALAD_USE_DEFAULT_GIT is set, the system git is at least v2.22.0,
and the bundled git is below v2.22.0.  In this case, 'git annex add
empty-subrepo' will have the old and undesirable behavior of adding
the file to the parent repo.

Avoid this by dropping the unnecessary sub-repository paths from the
'git annex add' call.