TST+BF: travis: Fix exit code logic for upstream git run
tools/ci/install-latest-git.sh exits with 100 when the upstream Git
version matches the bundled one.  In this case, we should exit with a
status of 0.  For other errors, we should exit with a non-zero status.
And if all goes well and there is a newer upstream Git, we shouldn't
exit, continuing with the test run.

The logic to handle this, though, is faulty.  Due to an operator
precedence mistake, we exit with 0 not only when install-latest-git.sh
exits with 100 but also when it exits with 0.  So we get a passing
build, but the tests are never executed.

Another issue with this logic is that we didn't exit early if
install-latest-git.sh exited with something other than 0 or 100 (e.g.,
with 1 because it couldn't fine the bundled git version).

Correct the logic to match the intended behavior.