NF: rerun: Add --explicit flag
run() will execute a command in a dirty working tree if --explicit is
passed, but rerun() requires a clean state.  Add an --explicit flag to
rerun() that makes all the underlying run calls operate with
explicit=True.

In some cases, this doesn't play well with --since (because it applies
to all commits, and they all may not have the appropriate outputs
specified) or with --onto because the Git operations we
perform (particularly checkout and cherry-pick) may fail due to
conflicts with the working tree modifications.

Instead of (or in addition to) adding --explicit to rerun, we could
store the explicit setting in the run record itself.  This information
isn't essential to the run, but it would allow us to selectively rerun
a subset of the commits in a range with explicit=True.  That's more
involved and intrusive, though, so let's hold off on that until there
is actually a demand for it.

Closes #3074.