BF: annexrepo: Adjust overly selective is_special_annex_remote()
To decide if a remote is an annex special remote, we check whether
'annex-externaltype' or 'annex-webdav' is configured for the remote.
This mis-classifies nearly all internal special remotes (anything but
webdav) as an ordinary remote because those don't have
'annex-externaltype' configured.

There's no common annex- option for special remotes.  As the
definition of annex's findSpecialRemotes shows, each special remote is
expected to have at least an annex-TYPE option.  We could keep a list
of known internal special remotes and look for that or externaltype,
but the list of internal special remotes is a moving target.

findSpecialRemotes' documentation mentions that "special remotes don't
have a configured url", so let's instead identify them based on _not_
having a URL and having some annex- option aside from annex-uuid and
annex-ignore, two particularly common non-special-remote options.

Given that all valid Git remotes have a configured URL, this should be
a reliable classification for correctly configured remotes.  It's of
course possible to get mis-classifications if the user either removes
or adds a URL to a remote's section.

Another option would be to rely on get_special_remotes(), which calls
`git cat-file git-annex:remote.log` to get the special remotes, but
that'd be a change in behavior because it also considers remotes that
aren't enabled.  It'd also be slower:

    # This patch
    % python -m timeit -s "from datalad.support.annexrepo import AnnexRepo; ar = AnnexRepo('.')"
      "ar.is_special_annex_remote('origin')"
    10000 loops, best of 3: 143 usec per loop

    # Using get_special_remotes() in is_special_annex_remote()
    % python -m timeit -s "from datalad.support.annexrepo import AnnexRepo; ar = AnnexRepo('.')"
      "ar.is_special_annex_remote('origin')"
    100 loops, best of 3: 8.82 msec per loop

Fixes #3497.