BF: subdatasets: Fix recursion when path is specified
_parse_git_submodules() returns early without yielding submodules if a
dataset's path does not equal or contain a supplied path.  This is
problematic when operating recursively because, if a directory
containing more than one level of subdatasets is given, the path is
the parent---not child---of the dataset, but it should still be
included in the output.  Add this condition to
_parse_git_submodules().

Fixes #3509.