ENH: annexrepo: Provide info about non-json from --json commands
When processing --json output of annex commands, we silently drop
lines that don't start with "{".  If we receive unexpected annex
output, this can lead to hard-to-troubleshoot bugs, such as the one
described in the next commit.

Based on reviewing commit messages, it's not clear if this is just to
drop empty lines or whether some commands are known to put non-json
output on stdout.  If non-json lines are detected, let's raise an
exception if no other valid json output was received and issue a
warning otherwise.  If it turns out that this warning happens in
common scenarios, we should demote it to debug().