MV: Move get_dataset_pwds back to run.py
This function was moved to utils.py to avoid duplicating some logic,
but it is no longer used anywhere outside of run and run wrappers.  To
update get_dataset_pwds() for the new path handling (gh-3435), it will
need to use distribution.dataset.Dataset, which doesn't seem
appropriate to import into utils.py, so let's move the helper back to
run.py.  We can move it somewhere else if the need arises.