BF: run: Update for new path resolution
Following the changes in gh-3435, teach run() that a specified dataset
might not be a dataset instance and that paths should be relative to
the dataset only if an instance is given.

The main change in behavior then is exposed when using --dataset with
a run call from the command line.  This can lead to some pretty weird
cases (e.g., the directory in which the command is executed can be
outside the dataset where the record is made).  But using 'run
--dataset ...'  from the command line isn't typical or (arguably)
recommended, so these cases probably aren't something to guard
against.