BF: Update more commands for EnsureDataset change
EnsureDataset no longer converts a string to a dataset instance as of
fc25ac55d (gh-3435).  This causes issues in commands that assume the
dataset argument, if specified, is a dataset instance.  A few spots
have already been updated (9af316f3d, bba278fe6, f38d72c68).  Fix what
are hopefully the only remaining issues, aside from those with run(),
which will be dealt with separately.

Note that most of the commands were checked with a combination of a
command line call with `-d .` and looking at the code for the dataset
handling, so it is likely that a couple of issues have been
overlooked, but hopefully the bulk of the problems are resolved at
this point.