TST(win): run: Avoid file_has_content() in path semantics test
The file_has_content() check fails on AppVeyor:

  https://ci.appveyor.com/project/mih/datalad/builds/25893947

The same failure can be triggered with

  ./tools/eval_under_testloopfs python -m nose -vs \
  datalad/core/local/tests/test_run.py:test_run_path_semantics

The root issue seems to be that, within GitRepo.get_changed_files(),
repo.git.diff() incorrectly reports the files as modified (dedicated
issue to follow).

The file_has_content() checks aren't essential for the purpose of this
test, so avoid them.