RF: repodates: Avoid calling cat-file through GitPython
It'd be good to replace these cat-file's with a batch call (like is
done elsewhere in gitrepo.py), but doing that would require more
thought because in this case the content doesn't have a predictable
number of lines.

Use _git_custom_command() for the reasons given in the last commit of
this series.