ENH: gitrepo: Add get_revisions() replacement for .git.rev_list()
Many of the remaining GitPython calls are to git.rev_list().  Add a
GitRepo method that covers most of these cases.

Note that new lines are not allowed in the format.  We could lift this
restriction by appending something like "%x00%x00marker%x00" to the
format (see the approach in -metalad's runprov.py) and adding special
processing, but none of the current callers need that flexibility, and
it is likely to be slower, so avoid it for now.