RF: Remove all remaining git.cmd.Git usage outside of repo classes
Replace these instances with one-off calls to _git_custom_command().
Ideally we wouldn't use _git_custom_command() because its name
indicates that it should be considered internal to gitrepo.py, but we
use it in plenty of other places already, so that ship has sailed, and
the only alternative at the moment is committing to a full-blown
GitRepo method.