RF: create: Let subdatasets fall through to dataset collision check
If create() is called with an existing subdataset as the path, we
abort saying that the subdataset is content in the parent dataset.
Although this is true, let's restrict this check to non-dataset
content (e.g., <existing subds>/<file>) and have the subdataset case
fall through to the "is subdataset?" check.  This will make it easier
to teach 'datalad create --force <existing subds>' to ignore the "is
subdataset?" check.

Re: #3549