BF: addurls: Fix construction of nested subpaths
As get_subpaths() moves across the path, it builds a list of subpaths.
Each subpath is the *extension* of the one before it.  For example,
"a//b//c" is broken into ["a", "a/b"].  The logic for joining the
subpaths is flawed because *all* previous subpaths in the list, not
just the last one, are joined together at the beginning of the next
subpath.  This leads to subpaths with duplicated elements if
get_subpaths() is called with a path that has more than two "//".

Use only the last subpath in the list to construct the start of the
new subpath.