RF: gitrepo: Rewrite get_submodules() to avoid GitPython
In addition to keeping with our general direction of moving away from
GitPython, this avoids an unresolved issue in GitPython's handling of
submodules [0].

As documented by the new test, there is a change in behavior when
get_submodules() is called in a repository that doesn't have a commit
checked out: it now returns any registered submodules instead of an
empty list.  This is consistent with the behavior of 'git submodule'
and 'datalad subdataset', and there doesn't seem to be an obvious
reason not to support it.

[0]: https://github.com/datalad/datalad/pull/3508#issuecomment-508462763