BF: compare annex version purely by the version based on the date
Fix first appeared in  7.20190708+git36-g32d526164-1 in NeuroDebian
but conda carried 7.20190708-gbc74eb1 which managed to not compare
"before" the +git36.
Actual fix was in 7.20190708-32-g7be690f32 from Jul 16, so let"s
just compare by the date at the cost of skipping on some neurodebians
for now.  After all it is just a regression test for annex itself