BF: addurls: Don't assume that the stream has rows
extract() fails with an index error if the stream is valid JSON or CSV
but lacks any rows (i.e. for JSON an empty list or for CSV a
header-only file).  Update extract() to issue a warning and return an
empty list of rows.

Re: https://github.com/templateflow/datalad-osf/issues/1
Closes #3577.