BF(py2): interface: Don't assume that results are unicode
We call text_type() on the result's `path` because it may come in as a
pathlib object.  Ideally we'd work with unicode internally within IO
boundaries, but when default_result_renderer() gets the results,
_process_results() has already encoded unicode strings.  This means
our handling fails with non-ascii paths on Python 2:

    $ datalad create Δ0
    [ERROR ] 'ascii' codec can't decode byte 0xce in position 16:
    ordinal not in range(128)
    [utils.py:default_result_renderer:506] (UnicodeDecodeError)

Avoid the above error by ensuring we use bytes for `path` on Python 2.

Note that the added test is oddly parametrized with one case because
the next commit will add another case.

Re: #3597