BF: sshrun: Add -4 and -6 for full Git compatibility
When a command is set for GIT_SSH_COMMAND and GIT_SSH_VARIANT is ssh,
Git expects the command to support the following parameters:

    [-p port] [-4] [-6] [-o option] [username@]host command

With the last commit, we support all but -4 and -6.  It seems unlikely
that these flags would cause problems in the wild because they are
specified via git push/fetch/pull, and we don't use those internally
or expose them as options in any datalad command.  But given we
advertise sshrun() as a replacement for GIT_SSH_COMMAND, we should
support all of the expected options.