TST: run_procedure: Test more arguments that need quoting
Unlike the spaces check in the last commit, this would fail before the
recent the shlex_quote() changes because the arguments contain quotes.

This very likely needs to be skipped on windows, where our quoting is
a no-op, but I'm leaving it as is in this commit to see how the
AppVeyor run fails.