ENH: annexrepo: Add check_direct_mode_support method
Using a class method here is a bit ugly.  The main motivation of doing
this---instead of adding a property or regular method to AnnexRepo or
adding a top-level helper to annexrepo.py---is to allow the method to
be called without instantiating the class while still sharing this
information with instantiated objects.