TST+RF: test_annexrepo: Split up a test
test_annex_version_handling patches our information on git annex so
that its version matches the minimum requirement, its version is
outdated, or annex is unavailable entirely.

The first scenario will interfere with the logic in
AnnexRepo.__init__() that decides whether the git-annex version
supports direct mode, so we need to skip the test if we're running a
git-annex version that doesn't support direct mode and
datalad.repo.direct is true.  Split this part off so that we can more
easily skip it without skipping the other scenarios.