BF+TST: Avoid leaking patched git-annex version
Several of the tests patch the git-annex version.  One of these,
test_annex_version_comparison, leaks state because it doesn't return
AnnexRepo.git_annex_version back to its original value.

Add a helper that takes care of this, and use it in all the spots that
temporarily override the git-annex version.