TST: annexrepo: Rework test to use check_repository_versions()
When testing that AnnexRepo(..., version=6) has the intended effect,
test_repo_version() asserts that we expect either a v6 or v7 repo
because more recent versions of git-annex auto-upgrade v6 to v7.  This
workaround requires us to keep up with the latest git-annex because v7
may start to auto-upgrade to a later version.

Instead determine which version the repository will end up at by
taking the closest supported version that git-annex reports.