ENH: annexrepo: Add check_repository_versions method
Provide a method for getting the supported and upgradable repository
versions.  In the current release of git-annex (7.20190819), the
supported versions will be 5 and 7 and the upgradable versions will be
0 through 6 (on non-Windows systems).  In an upcoming release, 5 will
likely be dropped from the supported versions [*].

This method will allow us to update code (particularly tests) that
assume v5 is the default supported version.

Use a classmethod for the same reason described in 66cf13b4b (ENH:
annexrepo: Add check_direct_mode_support method, 2019-08-29).

[*] https://git-annex.branchable.com/todo/v7_path_toward_default/

Re: #3628