ENH: annexrepo: Add check_direct_mode_support method, again
This is a partial cherry pick of 66cf13b4b (ENH: annexrepo: Add
check_direct_mode_support method, 2019-08-29).  When the 0.11.x
series (gh-3631) that contained it was merged in a19e5c88a, I dropped
the method because it had no callers.  But it turns out that I must've
botched my local testing with the unreleased annex and part of
test_direct_cfg does needs to be skipped.

0.11.x calls check_direct_mode_support() in __init__() to ensure that
the supports_direct_mode attribute is set to a non-None value, but
don't bother doing that here given that the only place in master that
needs to check direct mode support is one test.