BF: cmd: Don't leave unread lines on stream when logging online
When the Runner is called with log_online=True and we're logging
either stdout or stderr, the readline() call for the stream happens
under a `while proc.poll() is None`.  After that block, we may have
uncaptured output, either because the process was already finished by
the time we got to the proc.poll() condition or because each iteration
reads only a single line.  To collect this output, communicate() is
called.

However, communicate() may return an empty string even when there is
uncaptured output and appears to reliably do so if

  (1) we've called readline() on the stream
  (2) both stdout and stderr are set to subprocess.PIPE
  (3) we're running under Python 3

The second condition likely involves the fact that subprocess.py has
an optimization to avoid using threads if the number of pipes is below
two.

For example, try

    import subprocess as sp
    p = sp.Popen(["printf", "one\ntwo\n"], stdout=sp.PIPE, stderr=sp.PIPE)
    print(p.stdout.readline())
    print(p.communicate())

"two\n" is not included in the output returned by communicate().

To work around this, gather these remaining lines with a direct read()
call rather than relying on communicate() when we're using a pipe for
bother stdout and stderr.