DOC: Orthogonalize docs with handbook content
Concept: All docs remaining in here are technical and developer-focused.
The handbook has acquired all previously content use case descriptions,
and is now a complete and more comprehensive replacement for any
user/usage-focused information.

This commit strips away anything that is duplicate, while keeping
executable code that can still serve a purpose in the test battery.

In a future pass through the docs, and after the structure and content
of the handbook has matured a bit more, the content can be further
refined and more (and more useful) cross-references can be added.