BF: Supposedly cross-platform handling of local ssh urls
Turns local absolute paths into localhost ssh urls that git can understand
on Windows as well (with an OpenSSH server that uses bash as its default
login shell -- yes on windows).

None of this has implications on real functionality, it just tries to
get the tests to not stumble upon themselves.

Ultimately, it doesn't fully address the issue, because GitPy fails
inside for a yet to be determined reason. See
https://github.com/datalad/datalad/pull/3638 for more info.