RF: update() no longer uses AnnotatePaths
This also introduce slight changes in the API semantics, by removing
needless flexibility (ie. specifying a to-be-updated dataset either
via -d or as a path). Multi-seed recursive operation is also no longer
supported, in line with previous changes.

The PATH arg is now only used to constrain the updating of subdatasets
in recursive mode.

ping gh-3368