BF(v7): gitrepo: Avoid adding files to annex
With v6+ repositories, 'git add' adds the file to annex unless
annex.largefiles specifies that the file should go to git.  Typically
an annex repository will be represented as an AnnexRepo and files can
be added to git with `.add(..., git=True)`.  However, instantiating an
existing annex repo as a GitRepo leads to the surprising behavior that
GitRepo.add() adds the file to annex.  Configure annex.largefiles so
that GitRepo.add() always adds files to git.