DOC+TST: gitrepo: Document get_tags() order change from eed2ca6f6
Prior to the for_each_ref_() series (gh-3705), get_tags() returned
tags sorted by the committer date.  For annotated tags, this means
they were sorted by the object pointed to rather than the tagger date.
Following gh-3705, the sorted value is by for-each-ref's "creatordate"
field (which exists so that lightweight and annotated tags can be
sorted together), and annotated tags are now sorted based on the
tagger date.

This change seems harmless enough.  No current callers in the code
base rely on annotated tags being sorted by committer date, and it
seems unlikely that outside callers rely on this behavior.  But it
does mean that test_get_tags needs to patch GIT_COMMITTER_DATE when
creating annotated tags so that the tagger date is not affected by the
environment of the test run.

If we really want to preserve the old behavior, note that using
"committerdate" (as originally done by gh-3705's bb17d0ede) is not
correct.  It will sort lightweight tags by their committer date and
leave annotated tags sorted alphabetically at the top of the list.
Instead we'd have to use a more involved approach, such as including
both "committerdate" and the dereferenced "*committerdate" as fields
and sorting by whichever value is non-empty for a given entry.