TST: run: Work around intermittent failure
In the last couple of weeks, test_run_inputs_outputs has been failing
intermittently with a `git annex get` lock error.  While the failure
seems specific for this tests (down to the same spot), the source of
the problem isn't clear.  The test and underlying code in run.py
haven't been modified for a while, so there are no obvious culprits on
our side.  I haven't had any luck triggering the failure by running
the tests 30x locally using git-annex 7.20190912.

Let's see if we can work around this failure with a sleep.  Fixing the
error would of course be better, but until we can make progress with
that (e.g., a reliable recipe to trigger it locally), having a known
failure pop up intermittently is problematic because it makes it more
likely that unknown/relevant failures will be overlooked.

Re: #3653